Correções finais

+++ Itinera-EJB/src/main/java/br/com/itinera/persistencia/OrdemColetaDAO.java
@@ -1,8 +1,8 @@
-
 package br.com.itinera.persistencia;
 
 import br.com.itinera.modelo.Motorista;
 import br.com.itinera.modelo.OrdemColeta;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import javax.ejb.Stateless;
@@ -15,59 +15,59 @@ import javax.persistence.PersistenceContext;
  */
 @Stateless
 public class OrdemColetaDAO {
-    
+
     @PersistenceContext
     private EntityManager em;
-    
-    public List<OrdemColeta> listar(){
+
+    public List<OrdemColeta> listar() {
         return em.createNamedQuery("OrdemColeta.findAll").getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodoENotaFiscal(Date inicio,Date fim,Integer nota){
+
+    public List<OrdemColeta> buscarPorPeriodoENotaFiscal(Date inicio, Date fim, BigDecimal nota) {
         return em.createNamedQuery("OrdemColeta.findByPeriodoInicioFimNF")
-               .setParameter("inicio", inicio)
-               .setParameter("fim", fim)
-               .setParameter("numeroNota", nota)
-               .getResultList();
+                .setParameter("inicio", inicio)
+                .setParameter("fim", fim)
+                .setParameter("numeroNota", nota)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodo(Date inicio,Date fim){
+
+    public List<OrdemColeta> buscarPorPeriodo(Date inicio, Date fim) {
         return em.createNamedQuery("OrdemColeta.findByPeriodo")
-               .setParameter("inicio", inicio)
-               .setParameter("fim", fim)
-               .getResultList();
+                .setParameter("inicio", inicio)
+                .setParameter("fim", fim)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodoInicio(Date inicio){
+
+    public List<OrdemColeta> buscarPorPeriodoInicio(Date inicio) {
         return em.createNamedQuery("OrdemColeta.findByPeriodoInicio")
-               .setParameter("inicio", inicio)
-               .getResultList();
+                .setParameter("inicio", inicio)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodoFim(Date fim){
+
+    public List<OrdemColeta> buscarPorPeriodoFim(Date fim) {
         return em.createNamedQuery("OrdemColeta.findByPeriodoFim")
-               .setParameter("fim", fim)
-               .getResultList();
+                .setParameter("fim", fim)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorNotaFiscal(Integer nota){
+
+    public List<OrdemColeta> buscarPorNotaFiscal(BigDecimal nota) {
         return em.createNamedQuery("OrdemColeta.findByNumeroNota")
-               .setParameter("numeroNota", nota)
-               .getResultList();
+                .setParameter("numeroNota", nota)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodoInicioNotaFiscal(Date inicio,Integer nota){
+
+    public List<OrdemColeta> buscarPorPeriodoInicioNotaFiscal(Date inicio, BigDecimal nota) {
         return em.createNamedQuery("OrdemColeta.findByPeriodoInicioNF")
-               .setParameter("inicio", inicio)
-               .setParameter("numeroNota", nota)
-               .getResultList();
+                .setParameter("inicio", inicio)
+                .setParameter("numeroNota", nota)
+                .getResultList();
     }
-    
-    public List<OrdemColeta> buscarPorPeriodoFimNotaFiscal(Date fim, Integer nota){
+
+    public List<OrdemColeta> buscarPorPeriodoFimNotaFiscal(Date fim, BigDecimal nota) {
         return em.createNamedQuery("OrdemColeta.findByPeriodoFimNF")
-               .setParameter("fim", fim)
-               .setParameter("numeroNota", nota)
-               .getResultList();
+                .setParameter("fim", fim)
+                .setParameter("numeroNota", nota)
+                .getResultList();
     }
 
     public List<OrdemColeta> buscarPorMotorista(Motorista motorista, Date dtInicioFiltro, Date dtFimFiltro) {
@@ -77,18 +77,18 @@ public class OrdemColetaDAO {
                 .setParameter("dtFim", dtFimFiltro)
                 .getResultList();
     }
-    
-    public void inserir(OrdemColeta ordem){
+
+    public void inserir(OrdemColeta ordem) {
         em.persist(ordem);
     }
-    
-    public void alterar(OrdemColeta ordem){
+
+    public void alterar(OrdemColeta ordem) {
         em.merge(ordem);
     }
-    
-    public void excluir(OrdemColeta ordem){
+
+    public void excluir(OrdemColeta ordem) {
         OrdemColeta remover = em.merge(ordem);
         em.remove(remover);
     }
-    
+
 }
